network: add system_info network activator cloud.cfg overrides
Support overriding network activators in cloud.cfg system_info
on disk.

Default cloud-init activators are used either during hot-plug or
for datasources that are detected during init-network stage when
basic networking has already been setup by the OS.a

Activators are discovered in the following priority order and
determined based on the presence of related network tools:

system_info:
  network:
    activators: [eni, netplan, network-manager, networkd]

On some systems where multiple network config tools are installed
it is necessary to override the priority order of detection to ensure
one activator is chosen over the other. This is done by providing
the a custom cloud config snippet in a /etc/cloud/cloud.cfg.d/*cfg.

LP: #1958377