sources/azure: add experimental support for preprovisioned os disks
Some pre-provisioning scenarios require that the VM be started and
shut back down as part of preparing the VM for future use.

When the PPS type is PreprovisionedOSDisk, report ready and wait for
host to shut down the VM.

Provisioning will resume normally on next boot, so do not write
a reported ready marker file.

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>