test: fix wireguard integration tests
Wireguard currently tests pass locally on LXD, however these tests fail
on any platform that requires ssh to connect to the instance. This
occurs since the process of setting up wireguard involves adding
interfaces, routes, and firewall rules, which results in an instance
that cannot be connected with as expected by pycloudlib.

This change "fixes" the test by stopping the wireguard service in
runcmd, which un-does the network changes that make the instance
inaccessible.