ifconfig parser: Commit current state of crime
The start of this is largely taken from netinfo.py
We could break out deeply nested ifs into their own smaller functions,
but for now the goal is just to just start dealing with with the
entirety of ifconfig's output.

The current status is: everything apart from IPv6 and bridge members.
Tests should follow, ASAP, to verify that any of this has actually any
merrit.