distros/manage_services: disable service
in order to support some modules on BSD, which could replace
base-services, we'll need to be able to disable those base-services.
This PR adds support for the "disable" command to 'systemct', BSD
'service', and OpenBSD 'rcctl', which currently suppor that.
For SysV init 'service', we follow the 'enable' path, and 'stop' the
service instead.

LP 1991024

Sponsored by: FreeBSD Foundation