OpenBSD: remove pkg_cmd_environ function
This environment variable is being set to the main mirror, which is
heavily under load. It also sets it to a deprecated and highly insecure
protocol.

OpenBSD's default behaviour is to read a list of mirrors from
installurl(5): https://man.openbsd.org/installurl.5

We can restore this behaviour by removing the contents of this function.
Note that the Equivalent NetBSD function makes vastly more sense.

LP 1991567

Sponsored by: FreeBSD Foundation