Fix last reported event possibly not being sent (#1796)
Fix last reported event possibly not being sent

The run of every cloud-init mode is wrapped in a reporting
context manager. The final flush of events before the process exits was
happening within the context manager, however, one final event is sent
when the context manager exits. Since this event isn't subject to
waiting for event flush, cloud-init can exit before this event gets
sent.

This commit fixes this issue and also adds logging of POST data when
POSTING to a URL.

LP: #1993836