net: skip duplicate mac check for netvsc nic and its VF (#1853)
When accelerated network is enabled on Azure, the host presents
two network interfaces with the same mac address to the VM: 
a synthetic nic (netvsc) and a VF nic, which is enslaved to the synthetic 
nic. 

The net module is already excluding slave nics when enumerating
interfaces. However, if cloud-init starts enumerating after the kernel
makes the VF visible to userspace, but before the enslaving has finished,
cloud-init will see two nics with duplicate mac.

This change will skip the duplicate mac error if one of the two nics 
with duplicate mac is a netvsc nic

LP: #1844191