docs: uprate analyze to performance page
As previously discussed, the Analyze page contains the content we want
to present under the "Performance" topic and should be uprated.

I have changed the title of the "Analyze" page to "Performance" so
that it shows correctly in the LHS menu. The anchors are the same,
and I have not changed the file name so as to avoid having to update
any links.

I took the opportunity to make some minor tidying edits to the text.
Hopefully these will add clarity and not change the intended meaning.