util: seedfrom URL append trailing forward slash on routes if absent
If NoCloud have a seedfrom base which doesn't contain a '%s' token,
or a query string, ensure the end of the base url has a trailing
forward slash.

Previously if NoCloud seedfrom=http://something didn't end with a
trailing forward slash, the URLs that would be queried would lack
path component separators:
 http://somethingmeta-data
 http://somethinguser-data
 http://somethingvendor-data

Allow util.read_seeded to append the trailing forward slash when
needed. This allows appending kernel commandline params to be as
brief as possible when providing `ds=nocloud-net;s=http://someurl`

read_seeded will not append a forward slash when '%s' is present
because that token is a placeholder in a string format which is
expected to be a fully resolvable well-formed URL path.