sources/azure: fix device driver matching for net config (#1914)
The ordering of NICs provided by IMDS may not match the order enumerated
by kernel.  As such, we do not have any guarantee that the nic we're
checking the driver for is the nic we think it is.

Instead of making any assumptions about how the nics are named, check
all interfaces by mac address.  If there is an interface using
"hv_netvsc", match against that.  If there is only one interface driver
that is not blacklisted, use that (in case it is not "hv_netvsc"), but
log a debug event.  If there are multiple hits, don't match against any
of the names and report a warning.

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>