distros/rhel.py: _read_hostname() missing strip on "hostname"
Debugging a related issue, I found that the determination of the
"previous-hostname" was wrong an RHEL-based distributions.

In distros/__init__.py: update_hostname() has this code:

```
            prev_hostname = self._read_hostname(prev_hostname_fn)
...
        (sys_fn, sys_hostname) = self._read_system_hostname()
...
        if sys_hostname and prev_hostname and sys_hostname != prev_hostname:
            LOG.info(
                "%s differs from %s, assuming user maintained hostname.",
                prev_hostname_fn,
                sys_fn,
            )
            return
```

sys_hostname was found to contain a trailing newline, while
prev_hostname did not. `sys_hostname != prev_hostname` is
therefore always true. The check effectively reduces to:

```
        if sys_hostname and prev_hostname:
```

Evidence can be found in the log file, as "assuming user maintained
hostname", even when this is not true.

prev_hostname is determined using _read_hostname().

sys_hostname is determined using _read_system_hostname() which
also calls _read_hostname().

However, only the prev_hostname is being stripped by _read_hostname():

```
        if self.uses_systemd() and filename.endswith("/previous-hostname"):
            return util.load_file(filename).strip()
        elif self.uses_systemd():
            (out, _err) = subp.subp(["hostname"])
            if len(out):
                return out
```

I was unable to determine when this problem was introduced. It may
exist as far back as since "self.uses_systemd()" has been returning true?

The fix is to add a call to strip on the result of the call to
command "hostname" the same way the strip is done on the file
containing the "previous-hostname".