machine-id: set to uninitilized to trigger regeneration on clones
In many cases, the removal or /etc/machine-id from systemd
environments results in an unbootable system.

Adapt cloud-init clean --machine-id to be systemd-aware.

systemd v. 237 and later will trigger /etc/machine-id regeneration
on next boot when the value of /etc/machine-id file is
'uninitialized\n'.

On systems without systemd, continue to remove /etc/machine-id when
cloud-init clean --machine-id is provided as there are triggers in
place to regenerate on absence of this file.

LP: #1999680