docs: Cosmetic improvements and styling (#2057)
Make the docs look more like the documentation produced on Ubuntu.com.

Make the default font size larger for accessibility.

Code literals are now the same size as regular font for the same reason,
but made slightly bolded to help pick them out from text (which should
help e.g. when commands are included as inline text and not a code block.

Give the sidebar headings a higher weight, for easier navigation.

For accessibility, change the weight and spacing between topics in the
side bars to help separate them and make it easier to tell where one link
ends and the next begins.