sources/azure: add networking check for all source PPS
There is a networking check in _poll_imds() which will attempt DHCP
again if networking is not up for source PPS.  With the previous
change to wait at least 20 minutes during provisioning for DHCP,
this additional round is not necessary.

Report failure if networking is not up for any mode of source PPS.
In practice, this is very unlikely as provisioning will typically
timeout within the 20 minute window the VM is attempting DHCP and
the source PPS VM will be deleted.

This fixes an (unobserved) issue where Savable PPS does not have
networking prior to _wait_for_all_nics_ready().

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>