Adapt DataSourceScaleway to upcoming IPv6 support (#2033)
Add support for IPv4 and/or IPv6 provided metadata. Upcoming
features will allow for Instances to be started with only IPv6
connectivity. The datasource must be able to fetch its metadata from
the following endpoints : 

* IPv4 on http://169.264.42.42/conf
* IPv6 on http://[fd00:42::42]/conf

This URL may eventually be overridden in the future by a DNS
resolvable URL defined in /etc/cloud/config.d/scaleway.conf

Add support for the configuration of one or many IP addresses when
provided by the `public_ips` metadata key.  In such configurations,
instances no longer have a `private_ip`. The previous configuration
method is kept for backward compatibility.

This commit also fixes a bug in the definition of ssh keys from tags.