azure/imds: retry fetching metadata up to 300 seconds
Instead of a fixed number of retries, allow up to 5 minutes to fetch
metadata from IMDS.  The current approach allows for up to 11 attempts
depending on the path.  Given the timeout setting, this can vary from
~11 seconds up to ~32 seconds depending on whether or not read/connection
timeouts are encountered.

Delaying boot on the rare occasion that IMDS is delayed is better than
ignoring the metadata as it ensures the VM is configured as expected.

This is a very conservative timeout and may be reduced in the future.

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>