net: Use `ip route add` to add static routes
Support for rfc3442 routes was added to cloud-init in 07b172. In this
commit, `ip route append` was used to add the route. This succeeds at adding
static routes, but its documented purpose is related to multipath routing. To
avoid confusion around the undocumented behavior being used, switch to
`ip route add`, which does the same thing. This will also simplify an
abstraction to support distros that do not use iproute2.