net: refactor hyper-v VF filtering and apply to get_interfaces()
Azure is introducing a new VF ("MANA") that will initially behave
similarly to mlx4/5 but cannot be denylisted in the same manner.
This is because the synthetic interface (hv_netvsc) will no longer
be required to function in the future which means we must
intelligently filter the VFs out instead of relying solely on the
driver name.

- Isolate filtering logic for Hyper-V's SR-IOV/VFs when used
  with synthetic hv_netvsc interfaces.

- Move the filter up to get_interfaces() from
  get_interfaces_by_mac_on_linux() to increase coverage of the
  filter. With this in place, we should be able to purge the
  "blacklist_drivers" across the codebase as it will no longer be
  necessary unless there are other paths to be considered.

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>