net: purge blacklist_drivers across net and azure
It was only used by Hyper-V which now has a filtering
mechanism that does not require the use of a denylist.

This exposed some issues with tests misspelling "hv_netvsc"
and using unmatched mac addresses.  This fixes those to work
with the current filter that does not rely on the driver name.

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>