schema: add network v1 schema definition and validation functions
Part 1 of 3: Define network-config-v1 schema which is inert and not
yet tied to either initial boot validation or the command:
cloud-init schema.

Part 2 will plumb this network schema validation to log warnings
Part 3 will extend the command cloud-init schema to process and raise
network validation errors from the commandline.

In this branch:
  - Use jsonschema.exceptions.best_match improving anyOf, allOf
    ValidationErrors with specific details avoiding the unhelpful
    message: "is not valid under any of the given schemas"
  - Add static schema-cloud-config-v1.json to represent supported
    network config version 1
  - validate_cloudconfig_schema and get_schema now accept a new param
    config_type to select either network-config cloud-config schemas