net/sysconfig: enable sysconfig renderer if network manager has ifcfg-rh plugin
Some distributions like RHEL does not have ifup and ifdown scripts that
traditionally handled ifcfg-eth* files. Instead RHEL uses network manager with
ifcfg-rh plugin to handle ifcfg scripts. Therefore, the sysconfig should check
for the existence of ifcfg-rh plugin in addition to checking for the existence
of ifup and ifdown scripts in order to determine if it can handle ifcfg files.
If either the plugin or ifup/ifdown scripts are present, sysconfig renderer can
be enabled.

fixes: https://github.com/canonical/cloud-init/issues/4131
RHBZ: 2194050

Signed-off-by: Ani Sinha <anisinha@redhat.com>