nm: generate ipv6 stateful dhcp config at par with sysconfig (#4115)
The sysconfig renderer sets the following in the ifcfg file for IPV6 stateful
DHCP configuration:

 BOOTPROTO = "dhcp"
 DHCPV6C = True
 IPV6INIT = True
 IPV6_AUTOCONF = False

This should result in
  [ipv6]
  method=dhcp

in the network manager generated keyfile as DHCPV6C is set and
IPV6_AUTOCONF is not set. Unfortunately the network manager renderer
deviates from this and generates:
  [ipv6]
  method=auto

in it's rendered keyfile. This change fixes this deviation and sets the
IPV6 dhcp stateful configuration in alignment with what is generated by the
sysconfig renderer.

RHBZ: 2207716

Signed-off-by: Ani Sinha <anisinha@redhat.com>