BSD: remove datasource_list from cloud.cfg template
This limited list was put into place, because in the past, something
wasn't working right. Whatever it was, seems to be resolved. Removing it
means BSDs have the same chance on all clouds like other distros do.

People can always create images that customize the selection, but the
out of box experience should be that cloud-init works as advertised on
all supported distros and all available clouds.

Sponsored by: The FreeBSD Foundation