gce: improve ephemeral fallback NIC selection
While setting up the ephemeral network config for fetching the IMD at
init-local timeframe on multi-NIC instances, fulfilling:

- All NICs without carrier flag or more than one NIC with carrier flag
- systemd's predictable interface names is enabled

net.find_fallback_nic could select a NIC that is not the primary NIC,
leaving the instance without access to the network, as on GCE, only the
primary NIC can talk to the IMDS.

At this point in time, there is yet not access to
{instance,vendor,user}-data. Thus, it is tricky to dynamically inject a
breadcrum pointing to the primary NIC.

Two actions have been taken to fix this situation:

1. Substitute eth0 with ens4 as the default primary NIC candidate on
Linux systems.

2. Try through the list of candidate NICs and use the first that can
reach the IMDS.

Note that none of these changes do change cloud-init's behavior as:

(1) only applies to instances with more than one NIC. (2) will
eventually find the correct NIC as opposed to leaving the instance
without network (previous behavior).