Set default renderer as sysconfig for centos/rhel
Currently, network manager is disabled on c9s and therefore sysconfig is used as
the primary renderer for network configuration. We do not want to change this
for c9s even when network-manager renderer is re-enabled as it would mean
a big behaviour change for cloud-init in the centos 9 stream.

This change bumps up the priority for sysconfig renderer so that it is used as
the primary renderer on c9s and other downstream distributions derived
from it. In the next major centos stream release, we may use network manager as
the default renderer and make changes accordingly.

RHBZ:2209349

Signed-off-by: Ani Sinha <anisinha@redhat.com>