setup: fix generation of init templates
if an init file ends in any of the letters [tmpl] these will be stripped
in addition to the .tmpl extension.

This is documented behaviour of the str.rstrip() method.
This PR fixes it by using os.path.splitext() (rather than
str.removesuffix which is only available in python 3.9+)

Sponsored by: The FreeBSD Foundation