instance-data: add system-info and features to combined-cloud-config
Some snap environments (subiquity live installer) have different
versions of cloudinit within the snap environment than is installed
on the host system. Invoking cloudinit.stages.Init().read_cfg on an
older version of cloudinit potentially breaks when the host environment
has a newer version of cloud-init because the obj.pkl can not be
deserialized on an older version of cloud-init.

To avoid having to invoke cloudinit.stages.Init().read_cfg() to get
system_info and default_user config, provide this config in:
 - /run/cloud-init/combined-cloud-config.json and
 - /run/cloud-init/instance-data-sensitive.json:merged_system_cfg key

Add a system_info key in combined-cloud-config.json as seen in
/etc/cloud/cloud.cfg*.

Also add a util.get_features function which will seed the "features"
key in combined-cloud-config.json. This also prevents the need for
importing cloudinit.features to read that value or overridden feature
values.