Improved merging guide for actual users
This page is tremendously unhelpful right now. As a user, I'm presented a deep-dive on the technical implementation of the Python merger, which is essentially maintainer, or at least vendor/integrator documentation. There's _way_ too much detail on how this feature works, as opposed to what this feature actually achieves. The unclear target audience is something I've seen in lots of other places, too, and it makes it really difficult to implement user data sections.  

This change changes the order of the sections, makes several of them more concise, and adds additional examples.

There's still room for improvement; I _think_ I know what the section [_"Specifying multiple types, and what this does"_](https://cloudinit.readthedocs.io/en/latest/reference/merging.html#specifying-multiple-types-and-what-this-does) means, but I'm still not sure what happens if I use the `merge_how` key in two files and merge them, or  what happens to strings inside dictionaries.