Flatten landscape-config params and use distro.manage_service
- use cloud.distro.manage_service to restart instead of direct subp

- Avoid trying to combine '--key="value"' and escaping the command
values directly in cloud-init. Allow subp to process positional
--key and value parameters separately and properly escape
white-space and quotes when needed.

This avoids landscape-config --computer-title="My name" actually
preserving double-quotes as part of the value in
/etc/landscape/client.conf
   computer_title = "My name"

The quoted values of computer_title or log_level prevent
landscape-config from parsing the file properly and actually
brokering registration of the node.