testing: Clear all LRU caches after each test
Some unit tests were missing necessary mocks because a previous test
had mocked the behavior and mock stuck around because of an LRU cache.
This means that if tests got run out of a specific order, they would
fail due to missing mocks. Because of this, a new autouse fixture was
added to automatically clear any lru cache after every test.

Additionally, updated unit tests accordingly so that they have the
necessary mocks, and convert `test_create_users.py` to pytest as
I was getting some transient errors from it.