lxd: install lxd from snap, not deb if absent in image
When lxd: cloud-config is provided and lxd command is not
present on the image, install lxd from snap instead of deb.

On Ubuntu Focal and later LXD is available as a snap, the
transitional lxd deb package was no longer available for
install as a deb package as of 22.04 (Jammy) and later.

All ubuntu cloud images contain LXD as snap within the
base image manifest. So, this logic generally should not
get triggered. But, errors in cloud images build process
generated some images without LXD and that triggered
the install_packages logic to fail in cloud-init integration
tests.