Revert else block in network manager loop
In testing, returning False if all network managers failed to retrieve
data caused the instance to re-run cloud-init on reboot.

In this case, the cached `instance-id` would be `iid-datasource-none`,
as DataSourceNone was chosen after DataSourceAkamai did not return
metadata.  However, on the next boot DataSourceAkamai would be asked for
the instance id and return the correct value from dmi data, causing the
cache to be invalided and cloud-init to re-run.

DataSourceAkamai populates metadata before making requests (limited to
only the instance id which can be relieved locally), which is enough to
populate the cache and prevent a reprovisioning.  This causes the
"region" field to not be populated, and the user data to be missed; it
seems better to require a user to clean/reboot to consume user data than
it does to have a service outage set up long-lived instances to possibly
rotate host keys and consume user data unexpectedly on a subsequent
reboot.