perf!: refactor GCE IMDS query to pull all data [broken]
In order to improve boot performance, this change refactors the GCE IMDS
metadata fetching code to pull all IMDS data in a single query and then
parse out the results rather the querying the IMDS for each key
cloud-init needs to fetch.

This commit contains broken unit tests and is being pushed to
facillitate discussion of the use of bytestrings in the unit test.