ec2: Support double encoded userdata (#4276)
ec2: Support double encoded userdata

The Amazon APIs expect userdata to be base64 encoded when passed as
input to e.g. RunInstances.  A number of tools, including the AWS CLI,
perform this base64 encoding implicitly, but it's common for users to
base64 encode the data prior to passing it to them. This results in two
layers of base64 encoding and effectively results in a failed EC2
launch.  This change adds the ability to decode the redundant layer of 
encoding.

Fixes https://github.com/amazonlinux/amazon-linux-2023/issues/401

Signed-off-by: Noah Meyerhans <nmeyerha@amazon.com>