analyze: fix (unexpected) timestamp parsing
In case of very unexpected timestamps, we pass them on to date.
Unfortunately, only GNU date is able to perform this level of deduction
(guess work).

Rework the code and tests to look for GNU date on non-Linux platforms.
Only fail, by throwing an Exception! if GNU date cannot be found.
add `timestampstr` to that exception, so we know why we're in this code
path to begin with.

Sponsored by: The FreeBSD Foundation

This patch (partially)

Fixes: GH-4333

Co-authored-by: Brett Holman <brett.holman@canonical.com>